move unit tests to folder /test/unit/
