+++ test/unit/fixtures/o-test/demos/src/flat/example-1.mustache
