+++ test/unit/fixtures/o-test/demos/src/flat/example-2.mustache
