+++ test/unit/fixtures/o-test/demos/src/partial1.mustache
