+++ test/unit/fixtures/o-test/demos/src/partials/partial2.mustache
