+++ test/unit/fixtures/o-test/src/js/require-text.txt
