+++ test/unit/fixtures/o-test/src/js/syntax-error.js
