+++ test/unit/helpers/files.test.js
@@ -7,10 +7,10 @@ const expect = require('expect.js');
 const fs = require('fs-extra');
 const path = require('path');
 
-const files = require('../../lib/helpers/files');
+const files = require('../../../lib/helpers/files');
 
 const obtPath = process.cwd();
-const oTestPath = 'test/fixtures/o-test';
+const oTestPath = 'test/unit/fixtures/o-test';
 const pathSuffix = '-file';
 const filesTestPath = path.resolve(obtPath, oTestPath + pathSuffix);
