+++ test/unit/tasks/demo-build.test.js
@@ -7,10 +7,10 @@ const fs = require('fs-extra');
 const path = require('path');
 const sinon = require('sinon');
 
-const demo = require('../../lib/tasks/demo-build');
+const demo = require('../../../lib/tasks/demo-build');
 
 const obtPath = process.cwd();
-const oTestPath = 'test/fixtures/o-test';
+const oTestPath = 'test/unit/fixtures/o-test';
 const pathSuffix = '-demo';
 const demoTestPath = path.resolve(obtPath, oTestPath + pathSuffix);
