+++ test/unit/tasks/install-npm.test.js
@@ -13,12 +13,12 @@ sinon.assert.expose(proclaim, {
 });
 
 const obtPath = process.cwd();
-const oTestPath = 'test/fixtures/o-test';
+const oTestPath = 'test/unit/fixtures/o-test';
 const pathSuffix = '-verify';
 const verifyTestPath = path.resolve(obtPath, oTestPath + pathSuffix);
 
-describe('install-bower', function () {
-	let bowerInstall;
+describe('npm-install', function () {
+	let npmInstall;
 	let Listr;
 	let listrInstance;
 
@@ -37,9 +37,9 @@ describe('install-bower', function () {
 
 		mockery.registerMock('listr', Listr);
 
-		mockery.registerAllowable('../../lib/tasks/install-bower');
+		mockery.registerAllowable('../../../lib/tasks/install-npm');
 
-		bowerInstall = require('../../lib/tasks/install-bower');
+		npmInstall = require('../../../lib/tasks/install-npm');
 
 		mockery.resetCache();
 
@@ -56,27 +56,27 @@ describe('install-bower', function () {
 	});
 
 	it('has a default title', () => {
-		proclaim.equal(bowerInstall().title, 'Installing Bower components');
+		proclaim.equal(npmInstall().title, 'Installing NPM components');
 	});
 
 	describe('skip', () => {
-		it('should return true if bower.json does not exist', function() {
-			return bowerInstall().skip()
+		it('should return true if package.json does not exist', function () {
+			return npmInstall().skip()
 				.then(skipped => {
 					proclaim.ok(skipped);
 				});
 		});
 
-		it('should return a helpful message if bower.json does not exist', function() {
-			return bowerInstall().skip()
+		it('should return a helpful message if package.json does not exist', function () {
+			return npmInstall().skip()
 				.then(skipped => {
-					proclaim.equal(skipped, 'No bower.json found.');
+					proclaim.equal(skipped, 'No package.json found.');
 				});
 		});
 
-		it('should return a falsey value if bower.json does exist', function() {
-			fs.writeFileSync('bower.json', '{}');
-			return bowerInstall().skip()
+		it('should return a falsey value if package.json does exist', function () {
+			fs.writeFileSync('package.json', '{}');
+			return npmInstall().skip()
 				.then(skipped => {
 					proclaim.notOk(skipped);
 				});
@@ -84,12 +84,13 @@ describe('install-bower', function () {
 	});
 
 	describe('task', function () {
-		it('should create Listr object with verify tasks', function() {
-			bowerInstall().task();
+		it('should create Listr object with npm install function', function () {
+			npmInstall().task();
 
 			proclaim.calledOnce(Listr);
 			proclaim.calledWithNew(Listr);
 			proclaim.isArray(Listr.firstCall.args[0]);
 		});
+		// TODO: Should assert that npm install function is passed outputStreams or should this be done via integration tests?
 	});
 });
